Add hourly support to timeseries
This helps support #32 and #42. We could support other intervals, as
well, with a simple addition to the processData method. However, right
now we only support "hourly" and "daily", and intentionally barf on
other intervals.

Note that hourly data gets big (30k+ points) for some data streams, so
it might make sense to implicitly or explicitly limit the amount of data
transferred (say, only in the last year, so ~9000k points). Or maybe a
point count. Anyways, something needs to be done because the data
volumes get unmanageable.